Pix2Pix_Project/requirements.txt
absl-py==0.11.0
astroid==2.4.2
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
importlib-metadata==3.3.0
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
mpmath==1.1.0
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.1.48
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
toml==0.10.2
tqdm==4.55.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
